removed borrow history pipe
