+++ components/AppContext.ts
@@ -109,8 +109,6 @@ import {
   checkUserSettingsLocalStorage$,
   saveUserSettingsLocalStorage$,
 } from 'features/userSettings/userSettingsLocal'
-import { createVaultHistory$ } from 'features/vaultHistory/vaultHistory'
-import { createVaultMultiplyHistory$ } from 'features/vaultHistory/vaultMultiplyHistory'
 import { createVaultsOverview$ } from 'features/vaultsOverview/vaultsOverview'
 import { isEqual, mapValues, memoize } from 'lodash'
 import { curry } from 'ramda'
@@ -161,6 +159,7 @@ import { BalanceInfo, createBalanceInfo$ } from '../features/shared/balanceInfo'
 import { createCheckVaultType$, VaultIdToTypeMapping } from '../features/shared/checkVaultType'
 import { jwtAuthSetupToken$ } from '../features/termsOfService/jwt'
 import { createTermsAcceptance$ } from '../features/termsOfService/termsAcceptance'
+import { createVaultHistory$ } from '../features/vaultHistory/vaultHistory'
 import { doGasEstimation, HasGasEstimation } from '../helpers/form'
 import { createProductCardsData$ } from '../helpers/productCards'
 
@@ -482,9 +481,9 @@ export function setupAppContext() {
   )
 
   const vaultHistory$ = memoize(curry(createVaultHistory$)(context$, onEveryBlock$, vault$))
-  const vaultMultiplyHistory$ = memoize(
-    curry(createVaultMultiplyHistory$)(context$, onEveryBlock$, vault$),
-  )
+  // const vaultHistory$ = memoize(
+  //   curry(createVaultMultiplyHistory$)(context$, onEveryBlock$, vault$),
+  // )
 
   pluginDevModeHelpers(txHelpers$, connectedContext$, proxyAddress$)
 
@@ -663,7 +662,7 @@ export function setupAppContext() {
         exchangeQuote$,
         addGasEstimation$,
         userSettings$,
-        vaultMultiplyHistory$,
+        vaultHistory$,
         saveVaultUsingApi$,
         automationTriggersData$,
         id,
@@ -700,7 +699,7 @@ export function setupAppContext() {
         psmExchangeQuote$,
         addGasEstimation$,
         getProportions$,
-        vaultMultiplyHistory$,
+        vaultHistory$,
         userSettings$,
         id,
       ),
@@ -786,7 +785,6 @@ export function setupAppContext() {
     automationTriggersData$,
     accountData$,
     vaultHistory$,
-    vaultMultiplyHistory$,
     collateralPrices$,
     termsAcceptance$,
     reclaimCollateral$,
