+++ features/multiply/manage/pipes/manageMultiplyVaultConditions.ts
@@ -610,9 +610,7 @@ export function applyManageVaultConditions(
   ] as ManageMultiplyVaultStage[]).some((s) => s === stage)
 
   const stopLossTriggered =
-    !!vaultHistory[1] &&
-    'triggerId' in vaultHistory[1] &&
-    vaultHistory[1].eventType === 'executed'
+    !!vaultHistory[1] && 'triggerId' in vaultHistory[1] && vaultHistory[1].eventType === 'executed'
 
   return {
     ...state,
